+++ lib/pages/dashboards/dashboards_list_page.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:lotti/blocs/dashboards/dashboards_page_cubit.dart';
+import 'package:lotti/widgets/app_bar/sliver_title_bar.dart';
 import 'package:lotti/widgets/dashboards/dashboards_app_bar.dart';
 import 'package:lotti/widgets/dashboards/dashboards_list.dart';
-import 'package:lotti/widgets/habits/habit_page_app_bar.dart';
 
 class DashboardsListPage extends StatelessWidget {
   const DashboardsListPage({super.key});
@@ -19,7 +19,10 @@ class DashboardsListPage extends StatelessWidget {
         body: SafeArea(
           child: CustomScrollView(
             slivers: <Widget>[
-              SliverTitleBar(localizations.navTabTitleInsights),
+              SliverTitleBar(
+                localizations.navTabTitleInsights,
+                pinned: true,
+              ),
               const DashboardsSliverAppBar(),
               const DashboardsList(),
             ],